requirements.txt
praw
groq
python-dotenv
schedule
streamlit